Rename variables
The variables that were named *_url did not actually refer to values
that were URLs (Universal Resource Locator, ie. "http://example.com/"),
but rather to file and directory paths. Changing the variable names to
something meaningful/relevant, such as *_dir for directories or *_path
for file paths, makes it a bit easier to work with the software.