Optimize filename lookup
The previous algorithm walked the entire docs directory for every link
in every file, which quite expectedly took a lot of time. This update
makes a dict of the filenames to their respective absolute paths, making
the lookup O(1) instead of O(N), where N is the amount of files.

This reduced the cumulative time spent in the plugin from 13 out of 19
seconds, to 0.24 out of 7.0 seconds, when running `mkdocs build` with
cProfile on unitystation/unitystation-wiki.