Refactorings and optimizations (#16)
* Rename variables

The variables that were named *_url did not actually refer to values
that were URLs (Universal Resource Locator, ie. "http://example.com/"),
but rather to file and directory paths. Changing the variable names to
something meaningful/relevant, such as *_dir for directories or *_path
for file paths, makes it a bit easier to work with the software.

* Refactor path math

The path calculations were a bit convoluted, so this simplifies them a bit.

* Use the logging library

This makes the plugin use python's standard logging system and mkdocs'
configuration of that, as done in mkdocs/mkdocs#2082,
fralau/mkdocs-mermaid2-plugin#18 and fralau/mkdocs_macros_plugin#41.

* Optimize filename lookup

The previous algorithm walked the entire docs directory for every link
in every file, which quite expectedly took a lot of time. This update
makes a dict of the filenames to their respective absolute paths, making
the lookup O(1) instead of O(N), where N is the amount of files.

This reduced the cumulative time spent in the plugin from 13 out of 19
seconds, to 0.24 out of 7.0 seconds, when running `mkdocs build` with
cProfile on unitystation/unitystation-wiki.

* Simplify link generation

* Remove unused method parameter

* Fix hash anchor matching

The hash anchor group matched eagerly, which caused all kinds of errors
after the simplification in 4e6c276869e66bf, since it would match past
the actual end parenthesis of the link if it could.

https://github.com/midnightprioriem/mkdocs-autolinks-plugin/pull/5/files#r580470566